+++ EmbeddedPkg/Library/FdtLib/fdt_sw.c
@@ -81,7 +81,7 @@ static void *_fdt_grab_space(void *fdt, size_t len)
 	if ((offset + len < offset) || (offset + len > spaceleft))
 		return NULL;
 
-	fdt_set_size_dt_struct(fdt, offset + len);
+	fdt_set_size_dt_struct(fdt, offset + (int)len);
 	return _fdt_offset_ptr_w(fdt, offset);
 }
 
@@ -170,7 +170,7 @@ int fdt_finish_reservemap(void *fdt)
 int fdt_begin_node(void *fdt, const char *name)
 {
 	struct fdt_node_header *nh;
-	int namelen = strlen(name) + 1;
+	int namelen = (int)strlen(name) + 1;
 
 	FDT_SW_CHECK_HEADER(fdt);
 
@@ -202,17 +202,17 @@ static int _fdt_find_add_string(void *fdt, const char *s)
 	char *strtab = (char *)fdt + fdt_totalsize(fdt);
 	const char *p;
 	int strtabsize = fdt_size_dt_strings(fdt);
-	int len = strlen(s) + 1;
+	int len = (int)strlen(s) + 1;
 	int struct_top, offset;
 
 	p = _fdt_find_string(strtab - strtabsize, strtabsize, s);
 	if (p)
-		return p - strtab;
+		return (int)(p - strtab);
 
 	/* Add it */
 	offset = -strtabsize - len;
 	struct_top = fdt_off_dt_struct(fdt) + fdt_size_dt_struct(fdt);
-	if (fdt_totalsize(fdt) + offset < struct_top)
+	if ((int)fdt_totalsize(fdt) + offset < struct_top)
 		return 0; /* no more room :( */
 
 	memcpy(strtab + offset, s, len);
